+++ core/model/group.go
@@ -64,6 +64,12 @@ type Group struct {
 	SyncEndTime   *time.Time `json:"sync_end_time" bson:"sync_end_time"`
 } // @name Group
 
+// GetGroupMembershipsResponse response
+type GetGroupMembershipsResponse struct {
+	Title  string `json:”title`
+	Status string `json:"status"`
+} // @name GetGroupMembershipsResponse
+
 // ApplyLegacyMembership applies legacy membership to the group for backward compatibility
 func (gr *Group) ApplyLegacyMembership(membershipCollection MembershipCollection) {
 	var list []Member